fastapi/requirements.txt
fastapi==0.60.1
psycopg2-binary==2.8.6
